subocker/requirements.txt
pika
python-dotenv
requests